// src/const/buttons.enum.ts
export enum ButtonsText {
  Yes = 'Да',
  No = 'Нет',
  BecomePartner= 'Стать партнером',
  Review = 'Написать отзыв о продукте',
  About = 'О табаке "МОЛОДОСТЬ"',
  Taste = 'Вкусы',
  SocialNetwork = 'Социальные сети',
  WhereBuy = 'Где купить?',
  ReviewToOwner = 'Напрямую владельцу компании',
  ReviewToBot = 'Написать отзыв через бота',
  AboutPhilosophy = 'Философия',
  AboutCreators = 'Создатели',
  AboutProduct = 'О продукте',
  BecomePartnerToManager = 'Хочешь, чтобы с тобой связались?',
  BecomePartnerToBot = 'Хочешь написать сам?',

}

export enum ButtonsData {
  AgeYes = 'ageYes',
  AgeNo = 'ageNo',
  BecomePartner= 'menuBecomePartner',
  Review = 'review',
  About = 'about',
  Taste = 'taste',
  SocialNetwork = 'socialNetwork',
  WhereBuy = 'whereBuy',
  ReviewToOwner = 'reviewToOwner',
  ReviewToBot = 'reviewToBot',
  AboutPhilosophy = 'aboutPhilosophy',
  AboutCreators = 'aboutCreators',
  AboutProduct = 'aboutProduct',
  BecomePartnerToManager= 'becomePartnerToManager',
  BecomePartnerToBot = 'becomePartnerToBot',
}